refactor: AI 租户字段统一 Long 并补齐写操作规范
- AI 实体 5 个改继承 TenantBaseEntity（Long tenantId），AiMessage/AiDocument 类型改 Long
- currentTenantId() 全部收敛 Long，消除 Integer/Long 反复转换
- 新增 V7 迁移：ai_* 表 tenant_id INT 改 BIGINT
- 关系表 SysRoleMenu 等 5 个 @Data 改 @Getter @Setter
- SysApp/SysClient 写方法补 @Transactional，delete 业务化（接口声明+存在性校验）
- 补 @Log：kickout/bind/unbind/testSend；SysLog 权限改方法级
- SysMenuServiceImpl 租户静默 null 改 orElseThrow；License/SmsLogin 简化
- SKILL 补实体继承/写操作三件套/权限方法级/Resp 出参规范

+++ ypbin-admin-system/src/main/java/cn/ypbin/admin/system/ai/service/AiChatBizService.java
@@ -68,6 +68,6 @@ public interface AiChatBizService {
      *
      * <p>异步线程无请求上下文，租户 ID 由调用方（请求线程）显式传入。
      */
-    void saveAssistantMessageAsync(Long conversationId, Integer tenantId,
+    void saveAssistantMessageAsync(Long conversationId, Long tenantId,
             String content, int tokens);
 }
